\name{dcode-package}
\alias{dcode}
\docType{package}
\title{
	DCODE - Skeleton package to compile D 2.0 code inside the R 2.14.0 toolchain.
}

\description{
  Skeleton package to compile D 2.0 code inside the R 2.14.0 toolchain.
}

\details{
 None yet.
}

\author{ 
	Danny Arends \email{Danny.Arends@gmail.com}\cr
	Maintainer: Danny Arends \email{Danny.Arends@gmail.com}\cr
}

\keyword{package}
